\name{RanLetter-package}
\alias{RanLetter-package}
\alias{RanLetter}
\docType{package}
\title{
\packageTitle{RanLetter}
}
\description{
\packageDescription{RanLetter}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RanLetter}
\packageIndices{RanLetter}
~~ Randomly generates one letter from the English alphabet. ~~
}
\author{
\packageAuthor{Paul Davies}

Maintainer: \packageMaintainer{Paul Davies}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
